[CSL-2328] tests
Unfortunately, the hspec automatic Spec generator doesn't work now,
because our Specs have nonempty constraint contexts.
We should be able to rectify this in the near future, by getting rid of
the use of Data.Reflection to deliver these configuration values. The
only really difficult aspect of this is the slotting definitions, which
use protocol constants to get Bounded and Enum instances.